app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
choreographer==1.2.1
click==8.3.1
comm==0.2.3
cramjam==2.11.0
debugpy==1.8.18
decorator==5.2.1
executing==2.2.1
fastapi==0.124.4
fastparquet==2024.11.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
kaleido==1.2.0
logistro==2.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.13.0
nest-asyncio==1.6.0
numpy==2.3.5
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rich==14.2.0
simplejson==3.20.2
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
tornado==6.5.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.38.0
wcwidth==0.2.14
